requirements.txt
click
pillow
cairosvg
python-dotenv
